extract <OilProductionTable />
